+++ libs/nest/src/lib/health/health.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Req } from '@nestjs/common';
+import { Controller, Get } from '@nestjs/common';
 import { Public } from '../decorators';
 import { BaseWorker } from '../worker';
 import { HealthService } from './health.service';
@@ -15,14 +15,7 @@ export class HealthController extends BaseWorker {
    */
   @Get()
   @Public()
-  public async health(
-    @Req()
-    req: {
-      body: any;
-      headers: any;
-      query: any;
-    },
-  ) {
-    return this.service.health(req);
+  public async health() {
+    return this.service.health();
   }
 }